feat: Add repo iteration and simple job for updating LICENSE year (#2)
This adds code to index.ts that iterates over all of the user's repos,
and for each, executes a set of automations. One such automation is
implemented (although only very crudely). Its job is to look at the
last commit date, and see if it is newer than the year range specified
in the LICENSE file, and if so, create a pull request to update the
year range in LICENSE. There is absolutely no error handling or
checking for already existing branches or PRs yet. Hence, the CI step
to execute the script is disabled for safety.