feat: Implement repo config files (.graa.yml) and scoped logging
This patch adds reading of config files, of which there should be one
per repo. The file must be called '.graa.yml'. If no such file is
present in a repo, the repo will be skipped (making GRAA essentially
opt-in). A config file can specify an array of automations that should
be performed for the repo.

Additionally, a logging system is implemented that prefixes messages
with the repo full_name and potentially the automation id, if the
message is logged from an automation. This is done by hierarchically
creating loggers and passing them between functions.