feat: Implement merging of automation options for extended configs (#19)
This patch adds fully custom merging capabilities for automation option
objects. Previously, when an automation was specified with some set of
options A in a parent config (such as 'config:base'), and then again
with some set of options B in a child config (such as 'config:js-lib'),
only option set B would take effect. Now, each automation can specify
a reducer for merging its options to obtain a combined option set AB
which will be passed during automation execution. This is useful, for
example, for the 'files' automation. Specifying a file path twice will
override it, just as before, but specifying additional paths will not
replace the entire set of paths, but instead add to it, so that all
files, no matter where they are specified, will be processed.